backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
google-generativeai
passlib[bcrypt]
python-jose[cryptography]
fastapi-security
python-multipart
reportlab